# Setup default font manager after engine created, to improve startup performance