--- a/shell/platform/android/io/flutter/embedding/engine/FlutterJNI.java
+++ b/shell/platform/android/io/flutter/embedding/engine/FlutterJNI.java
@@ -106,6 +106,13 @@ public class FlutterJNI {
       @NonNull String engineCachesPath,
       long initTimeMillis);
 
+  /**
+   * Create the default font manager provided by SkFontMgr::RefDefault() which is a process-wide
+   * singleton owned by Skia. Note that, the first call to SkFontMgr::RefDefault() will take
+   * noticeable time, but later calls will return a reference to the preexisting font manager.
+   */
+  public static native void nativeCreateDefaultFontManager();
+
   // TODO(mattcarroll): add javadocs
   @UiThread
   public native boolean nativeGetIsSoftwareRenderingEnabled();