--- a/shell/platform/android/io/flutter/embedding/engine/loader/FlutterLoader.java
+++ b/shell/platform/android/io/flutter/embedding/engine/loader/FlutterLoader.java
@@ -144,16 +144,16 @@ public class FlutterLoader {
 
             System.loadLibrary("flutter");
 
-            // Pre-warm the default font manager as soon as possible on a background thread.
+            // Prefetch the default font manager as soon as possible on a background thread.
             // It helps to reduce time cost of engine setup that blocks the platform thread.
-            new Thread(
+            Executors.newSingleThreadExecutor()
+                .execute(
                     new Runnable() {
                       @Override
                       public void run() {
-                        FlutterJNI.nativeCreateDefaultFontManager();
+                        FlutterJNI.nativePrefetchDefaultFontManager();
                       }
-                    })
-                .start();
+                    });
 
             if (resourceExtractor != null) {
               resourceExtractor.waitForCompletion();