`Shell::Create` may gain 25~50% performance improvement by this pr ^_^

`SkFontMgr::RefDefault` is time-consuming, about 15ms in Pixel2, which is blocking Android's main thread during engine initialization. It is called by `FontCollection` while creating `Engine`, and I think we can just make that happen right after initialization instead of initializing.

before this pr: 
![image](https://user-images.githubusercontent.com/62057376/81392016-9d424a80-9150-11ea-8974-39339dbb09c8.png)

after this pr:
![image](https://user-images.githubusercontent.com/62057376/81392190-ded2f580-9150-11ea-996a-f85edd542bee.png)

Shown as the above figures,
Shell::Create    ~60ms ==> ~30ms
ShellSetupUISubsystem    ~25ms ==> ~10ms