@jason-simmons 2020-05-08
This is creating a race between the task that sets up the font manager and users of the engine that assume that the engine is fully initialized after it is created.

Most of the engine initialization work is not done on the Android main thread.  If the app is blocking the main thread during the call to `FlutterLoader.ensureInitializationComplete`, then consider using `ensureInitializationCompleteAsync` which will invoke a callback after engine setup finishes.