@liyuqian 2020-05-09
Cross reference https://github.com/flutter/engine/pull/18047 and https://github.com/flutter/engine/pull/17192. Those PRs also aim at speeding up startup time by making some work asynchronous. Most parts of shell and engine are initialized on other threads (e.g., UI thread), but it seems that the platform thread (Android's main thread) is currently blocked until all those initializations are finished (see, e.g., the `future` wait in https://github.com/flutter/engine/blob/9d8daf2383eadd70f0b539f8b6a199ab4815cc0c/shell/common/shell.cc#L167).

@jason-simmons : do we have a benchmark that verifies `ensureInitializationCompleteAsync` would allow Android main thread to be usable before the shell is fully initialized? If `Shell::Create` will block the platform thread before the `Shell` is fully initialized, I'm not sure whether `ensureInitializationCompleteAsync` could allow the app to do some platform thread work before that.