@jason-simmons 2020-05-09
`ensureInitializationCompleteAsync` will avoid blocking the main thread while waiting for one-time initialization work that is done by `FlutterLoader` in Java background threads.

But if the concern is the latency of `Shell::CreateShellOnPlatformThread`, then `ensureInitializationCompleteAsync` will not help.  Reducing blocking there would involve making shell construction asynchronous (as in the linked PRs).

The default font manager provided by `SkFontMgr::RefDefault()` is a process-wide singleton owned by Skia.  The first call to `SkFontMgr::RefDefault()` will take noticeable time, but later calls will return a reference to the preexisting font manager.  If your app allows it, it might help to prefetch the default font manager by creating a `FlutterEngine` at a time when the main thread is less loaded.

If we can't make shell creation async, then another potential workaround would be calling `SkFontMgr::RefDefault()` during the part of `FlutterLoader` startup that runs on a background thread.