@scutlight 2020-05-09
Since `SetupDefaultFontManager` task is posted inside `Shell::Setup`, it should be the very next task run after engine created, and users of `Shell` can still assume that it is ready to use once shell has been created.