@scutlight 2020-05-09
> If we can't make shell creation async, then another potential workaround would be calling `SkFontMgr::RefDefault()` during the part of `FlutterLoader` startup that runs on a background thread.
---------------------------------------------------------------------------
I've try this workaround by calling `SkFontMgr::RefDefault()` on a java background thread when libflutter.so loaded.  As the first call to `SkFontMgr::RefDefault()` takes time, it blocks the engine construction a while.
![image](https://user-images.githubusercontent.com/62057376/81461094-1e92ef00-91dc-11ea-80fd-31ab6b6e2533.png)