@scutlight 2020-05-09
One more thing, I suppose that the first call to `SkFontMgr::RefDefault()` is expensive on all platforms not just Android. It's suggested a solution inside shell instead of platform embedding.