@scutlight 2020-05-18
I've made a update.
As [flutter_boost](https://github.com/alibaba/flutter_boost) calls `executeDartEntrypoint` immediately after `FlutterEngine` was created, calling `SkFontMgr::RefDefault()` on a backgroud thread to pre-warm may be a seemly good workaround.
https://github.com/alibaba/flutter_boost/blob/bfb71ebd92b1c928d80881fe92a8fdff9a2552ad/android/src/main/java/com/idlefish/flutterboost/FlutterBoost.java#L158-L183