@liyuqian 2020-05-19
Just chatted with @jason-simmons offline. We think "there might be a race between the [deferred task][1] that calls `SkFontMgr::RefDefault` on the UI thread and any code that uses the engine to render text." Hence it's better to ensure that `SkFontMgr::RefDefault` is called before the engine is fully initialized.

Running `SkFontMgr::RefDefault` on another thread (not UI or platform thread) may be a way to parallelize the work and thus improve startup time as Jason suggested. Your screenshot in https://github.com/flutter/engine/pull/18225#issuecomment-626087798 seems to suggest that `SkFontMgr::RefDefault` still runs on the UI thread, and thus it's not parallelized. Does your modification yesterday in https://github.com/flutter/engine/pull/18225/commits/9788976ce7425f172edcddc997e02cde1e451cef make it parallel and improve the startup time? In addition to the Java solution that's probably Android -specific, you may also try something like `std::async(std::launch::async, [](){ SkFontMgr::RefDefault(); });` to make it cross-platform, and wait for its completion by the end of `Shell::Create` using `std::promise, std::future`.

[1]: https://github.com/flutter/engine/pull/18225/files#diff-a94514bcac1a44f912829825243144d6R546