@scutlight 2020-05-20
> Just chatted with @jason-simmons offline. We think "there might be a race between the [deferred task](https://github.com/flutter/engine/pull/18225/files#diff-a94514bcac1a44f912829825243144d6R546) that calls `SkFontMgr::RefDefault` on the UI thread and any code that uses the engine to render text." Hence it's better to ensure that `SkFontMgr::RefDefault` is called before the engine is fully initialized.

It is clear that "the setup task will run before any other UI thread tasks that might use fonts", refer to Jason's latest https://github.com/flutter/engine/pull/18225#pullrequestreview-414930491, and my previous https://github.com/flutter/engine/pull/18225#issuecomment-626086005

> Running `SkFontMgr::RefDefault` on another thread (not UI or platform thread) may be a way to parallelize the work and thus improve startup time as Jason suggested. Your screenshot in [#18225 (comment)](https://github.com/flutter/engine/pull/18225#issuecomment-626087798) seems to suggest that `SkFontMgr::RefDefault` still runs on the UI thread, and thus it's not parallelized. Does your modification yesterday in [9788976](https://github.com/flutter/engine/commit/9788976ce7425f172edcddc997e02cde1e451cef) make it parallel and improve the startup time? 

Sorry for confusing screenshot. It is parallelized. The UI thread is busy waiting until the singleton is ready due to the implementation of Skia's `SkOnce` called by `SkFontMgr::RefDefault`. That's also the reason why I move `SetupDefaultFontManager` out of engine creating.

> In addition to the Java solution that's probably Android -specific, you may also try something like `std::async(std::launch::async, [](){ SkFontMgr::RefDefault(); });` to make it cross-platform, and wait for its completion by the end of `Shell::Create` using `std::promise, std::future`.

Yeah, I agree that. Maybe we can put this `std::async` into `PerformInitializationTasks` of shell.cc to make it cross-platform. But I have no effort to cover tests to get the eventual result on other platforms. And more important, it's better to start `SkFontMgr::RefDefault()` as soon as possible because it really takes a long time to finish.

> BTW, how important is this improvement to your app, @scutlight? Is it "customer: critical" or "customer: blocker"? CC @zoeyfan for evaluating the importance from the customer's perspective.

Maybe not so emergent as critical, but really important since the whole engine startup is non-trivial and blocking the Android's main thread.